DBP-76 Changes for review, dummy modules for ansible-doc
