+++ onepwd/galaxy.yml
@@ -9,7 +9,7 @@ namespace: dbildungscloud
 name: onepwd
 
 # The version of the collection. Must be compatible with semantic versioning
-version: 2.0.3
+version: 2.1.0
 
 # The path to the Markdown (.md) readme file. This path is relative to the root of the collection
 readme: README.md
@@ -17,7 +17,7 @@ readme: README.md
 # A list of the collection's content authors. Can be just the name or in the format 'Full Name <email> (url)
 # @nicks:irc/im.site#channel'
 authors:
-- HPI Schul-Cloud <devops@hpi-schul-cloud.de>
+- HPI Schul-Cloud <devops@dbildungscloud.de>
 
 
 ### OPTIONAL but strongly recommended
